Booksmut/functions/hardcover-enricher/requirements.txt
functions-framework==3.*
requests==2.*
psycopg2-binary==2.*
google-cloud-secret-manager==2.*
